+++ src/service/schema.ts
@@ -1,17 +1,17 @@
 import { z } from 'zod'
 
-const ClientPermissionsEnum = z.enum([
-  'READ_ALL',
-  'READ_PUBLIC',
-  'READ_PRIVATE',
-  'READ_PROTECTED_{DOMAIN}',
-  'WRITE_ALL',
-  'WRITE_PUBLIC',
-  'WRITE_PRIVATE',
-  'WRITE_PROTECTED_{DOMAIN}',
-  'DATA_ADMIN',
-  'USER_ADMIN'
-])
+// export const ClientPermissionsEnum = z.enum([
+//   'READ_ALL',
+//   'READ_PUBLIC',
+//   'READ_PRIVATE',
+//   'READ_PROTECTED_{DOMAIN}',
+//   'WRITE_ALL',
+//   'WRITE_PUBLIC',
+//   'WRITE_PRIVATE',
+//   'WRITE_PROTECTED_{DOMAIN}',
+//   'DATA_ADMIN',
+//   'USER_ADMIN'
+// ])
 
 const UserTypeEnum = z.enum(['USER', 'CLIENT'])
 
@@ -20,4 +20,14 @@ export const SchemaUserCreate = z.object({
   name: z.string(),
   email: z.string().email(),
   permissions: z.array(z.string()).optional()
+})
+
+const keyValueTags = z.object({
+  key: z.string(),
+  value: z.string()
+})
+
+export const schemaCreateSchema = z.object({
+  keyValueTags: z.array(keyValueTags),
+  keyTags: z.array(z.string())
 })
